backend/requirements.txt
# Backend requirements
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
pydantic==2.5.3
pydantic-settings==2.1.0
cryptography==42.0.1
argon2-cffi==23.1.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pytest==7.4.4
httpx==0.26.0
